Emoji button input fixed
The emoji picker now allows you to add and emoji to the message wherever your cursor is located. I had to change the message box to an input element instead of a textarea element. The emoji picker is still bot showing the first three categories - will be working on this next.